Simplification de l'acquisition et du traitement depuis les trames can du code de la base roulante
Modification du code maitre pour tester le bon fonctionnement du can base roulante

+++ CodeBaseRoulanteOTA/lib/USE_FUNCTION/USE_FUNCTION.cpp
@@ -3,7 +3,7 @@
 
 #include <mat.h>
 
-int fusion_octet(int octet0, int octet1)
+int fusion_octet(int8_t octet0, int8_t octet1)
 {
 
     int16_t octet16 = (octet0 << 8) | octet1;
